requirements.txt
fastapi
uvicorn
confluent-kafka
faker
numpy
python-dotenv